fix: some problems

+++ smart-sso-starter/smart-sso-starter-base/src/main/java/openjoe/smart/sso/base/entity/ExpirationWrapper.java
@@ -28,7 +28,7 @@ public class ExpirationWrapper<T> {
         super();
         this.object = object;
         this.expiresIn = expiresIn;
-        this.expired = System.currentTimeMillis() + expiresIn * 1000;
+        this.expired = System.currentTimeMillis() + (long)expiresIn * 1000;
     }
 
     public T getObject() {
